fix(R.6): close merge-forward QA blockers
