+++ crates/atm/tests/send.rs
@@ -3,31 +3,13 @@
 mod support;
 
 use std::fs;
-use std::path::PathBuf;
+use std::ops::Deref;
 use std::process::Command;
 
 use crate::support::{
-    ROLE_TEAM_LEAD, TEST_LEAD, TEST_QA, TEST_RECIPIENT, TEST_RECIPIENT_ADDRESS, TEST_SENDER,
-    TEST_SENDER_ADDRESS, TEST_TEAM,
+    CliFixture, ROLE_TEAM_LEAD, TEST_LEAD, TEST_QA, TEST_RECIPIENT, TEST_RECIPIENT_ADDRESS,
+    TEST_SENDER, TEST_SENDER_ADDRESS, TEST_TEAM,
 };
-use atm_core::schema::{
-    AgentMember, MessageEnvelope, TeamConfig, hydrate_legacy_fields_from_metadata,
-};
-use serde_json::Value;
-
-fn parse_inbox_values(raw: &str) -> Vec<Value> {
-    if raw.trim().is_empty() {
-        return Vec::new();
-    }
-
-    match raw.chars().find(|ch| !ch.is_whitespace()) {
-        Some('[') => serde_json::from_str(raw).expect("json array"),
-        _ => raw
-            .lines()
-            .map(|line| serde_json::from_str(line).expect("json line"))
-            .collect(),
-    }
-}
 
 #[test]
 fn test_send_creates_inbox_file() {
@@ -166,7 +148,7 @@ fn test_send_requires_ack() {
     let atm_message_id = inbox[0].extra["metadata"]["atm"]["messageId"]
         .as_str()
         .expect("atm message id");
-    let workflow = fixture.workflow_state_contents(TEST_TEAM, TEST_RECIPIENT);
+    let workflow = fixture.workflow_state_contents_in_team(TEST_TEAM, TEST_RECIPIENT);
     assert!(
         workflow["messages"][format!("atm:{atm_message_id}")]["read"].is_null()
             || workflow["messages"][format!("atm:{atm_message_id}")]["read"] == false
@@ -1000,253 +982,18 @@ fn test_send_help_mentions_post_send_hook_config() {
     assert!(stdout.contains(".atm.toml"));
 }
 
-struct Fixture {
-    tempdir: tempfile::TempDir,
-}
+struct Fixture(CliFixture);
 
 impl Fixture {
     fn new(recipient: &str) -> Self {
-        let tempdir = tempfile::tempdir().expect("tempdir");
-        let fixture = Self { tempdir };
-        fixture.write_team_config(recipient);
-        fixture.warm_daemon();
-        fixture
+        Self(CliFixture::new_with_recipient(recipient))
     }
+}
 
-    fn run(&self, args: &[&str]) -> std::process::Output {
-        self.run_with_env(args, &[])
-    }
+impl Deref for Fixture {
+    type Target = CliFixture;
 
-    fn warm_daemon(&self) {
-        for attempt in 0..3 {
-            let output = self.run(&["read", "--all", "--no-mark", "--json"]);
-            if output.status.success() {
-                return;
-            }
-            assert!(
-                crate::support::is_daemon_start_transient(&output),
-                "stderr: {}",
-                self.stderr(&output)
-            );
-            if attempt < 2 {
-                std::thread::sleep(std::time::Duration::from_millis(100));
-            }
-        }
-        panic!("daemon warmup exhausted retries");
-    }
-
-    fn run_without_identity(&self, args: &[&str]) -> std::process::Output {
-        for attempt in 0..3 {
-            let mut first = Command::new(env!("CARGO_BIN_EXE_atm"));
-            let output =
-                crate::support::configure_atm_command(&mut first, self.tempdir.path(), None)
-                    .args(args)
-                    .current_dir(self.tempdir.path())
-                    .output()
-                    .unwrap_or_else(|error| {
-                        panic!(
-                            "atm without identity {:?} failed on attempt {attempt}: {error}",
-                            args
-                        )
-                    });
-            if output.status.success()
-                || !crate::support::is_daemon_start_transient(&output)
-                || attempt == 2
-            {
-                return output;
-            }
-
-            std::thread::sleep(std::time::Duration::from_millis(50));
-        }
-        unreachable!("send fixture identity-free retries should always return");
-    }
-
-    fn run_with_env(&self, args: &[&str], extra_env: &[(&str, &str)]) -> std::process::Output {
-        for attempt in 0..3 {
-            let mut first = Command::new(env!("CARGO_BIN_EXE_atm"));
-            crate::support::configure_atm_command(
-                &mut first,
-                self.tempdir.path(),
-                Some(TEST_SENDER),
-            )
-            .args(args)
-            .current_dir(self.tempdir.path());
-            for (key, value) in extra_env {
-                first.env(key, value);
-            }
-            let output = first.output().unwrap_or_else(|error| {
-                panic!("atm {:?} failed on attempt {attempt}: {error}", args)
-            });
-            if output.status.success()
-                || !crate::support::is_daemon_start_transient(&output)
-                || attempt == 2
-            {
-                return output;
-            }
-
-            std::thread::sleep(std::time::Duration::from_millis(50));
-        }
-        unreachable!("send fixture retries should always return");
-    }
-
-    fn write_team_config(&self, recipient: &str) {
-        self.write_team_config_for_team(TEST_TEAM, recipient);
-    }
-
-    fn write_team_config_for_team(&self, team: &str, recipient: &str) {
-        let team_dir = self.tempdir.path().join(".claude").join("teams").join(team);
-        fs::create_dir_all(&team_dir).expect("team dir");
-        let config = TeamConfig {
-            members: vec![AgentMember::with_name(recipient.parse().expect("agent"))],
-            ..Default::default()
-        };
-        fs::write(
-            team_dir.join("config.json"),
-            serde_json::to_vec(&config).expect("team config"),
-        )
-        .expect("write team config");
-    }
-
-    fn write_raw_team_config(&self, raw: &str) {
-        let team_dir = self
-            .tempdir
-            .path()
-            .join(".claude")
-            .join("teams")
-            .join(TEST_TEAM);
-        fs::create_dir_all(&team_dir).expect("team dir");
-        fs::write(team_dir.join("config.json"), raw).expect("write raw team config");
-    }
-
-    fn write_atm_config(&self, raw: &str) {
-        fs::write(self.tempdir.path().join(".atm.toml"), raw).expect("write .atm.toml");
-    }
-
-    fn inbox_path(&self, recipient: &str) -> std::path::PathBuf {
-        self.inbox_path_in_team(TEST_TEAM, recipient)
-    }
-
-    fn inbox_path_in_team(&self, team: &str, recipient: &str) -> std::path::PathBuf {
-        self.tempdir
-            .path()
-            .join(".claude")
-            .join("teams")
-            .join(team)
-            .join("inboxes")
-            .join(format!("{recipient}.json"))
-    }
-
-    fn write_inbox(&self, recipient: &str, messages: &[MessageEnvelope]) {
-        let inbox_path = self.inbox_path(recipient);
-        if let Some(parent) = inbox_path.parent() {
-            fs::create_dir_all(parent).expect("inbox dir");
-        }
-        let values: Vec<Value> = messages
-            .iter()
-            .map(|message| serde_json::to_value(message).expect("json value"))
-            .collect();
-        let raw = serde_json::to_string_pretty(&values).expect("json array");
-        fs::write(inbox_path, raw).expect("write inbox");
-    }
-
-    fn inbox_contents(&self, recipient: &str) -> Vec<MessageEnvelope> {
-        self.inbox_contents_in_team(TEST_TEAM, recipient)
-    }
-
-    fn inbox_json_lines(&self, recipient: &str) -> Vec<Value> {
-        self.inbox_json_lines_in_team(TEST_TEAM, recipient)
-    }
-
-    fn inbox_contents_in_team(&self, team: &str, recipient: &str) -> Vec<MessageEnvelope> {
-        let inbox_path = self.inbox_path_in_team(team, recipient);
-        let raw = fs::read_to_string(&inbox_path).expect("inbox contents");
-        parse_inbox_values(&raw)
-            .into_iter()
-            .map(|mut value| {
-                hydrate_legacy_fields_from_metadata(&mut value);
-                serde_json::from_value(value).expect("message envelope")
-            })
-            .collect()
-    }
-
-    fn inbox_json_lines_in_team(&self, team: &str, recipient: &str) -> Vec<Value> {
-        let inbox_path = self.inbox_path_in_team(team, recipient);
-        let raw = fs::read_to_string(&inbox_path).expect("inbox contents");
-        parse_inbox_values(&raw)
-    }
-
-    fn team_dir(&self) -> std::path::PathBuf {
-        self.tempdir
-            .path()
-            .join(".claude")
-            .join("teams")
-            .join(TEST_TEAM)
-    }
-
-    fn workflow_state_contents(&self, team: &str, agent: &str) -> Value {
-        let raw = fs::read_to_string(
-            self.tempdir
-                .path()
-                .join(".claude")
-                .join("teams")
-                .join(team)
-                .join(".atm-state")
-                .join("workflow")
-                .join(format!("{agent}.json")),
-        )
-        .expect("workflow state contents");
-        serde_json::from_str(&raw).expect("workflow json")
-    }
-
-    fn install_hook_fixture(&self, mode: &str) -> (PathBuf, PathBuf) {
-        let fixture_binary = PathBuf::from(env!("CARGO_BIN_EXE_atm_post_send_hook_fixture"));
-        let hook_dir = self.tempdir.path().join("bin");
-        fs::create_dir_all(&hook_dir).expect("hook dir");
-        let hook_path = hook_dir.join(fixture_binary.file_name().expect("hook binary filename"));
-        fs::copy(&fixture_binary, &hook_path).expect("copy hook fixture");
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-
-            let mut permissions = fs::metadata(&hook_path)
-                .expect("hook metadata")
-                .permissions();
-            permissions.set_mode(0o755);
-            fs::set_permissions(&hook_path, permissions).expect("hook permissions");
-        }
-        let payload_path = self.tempdir.path().join(format!("{mode}-payload.json"));
-        (
-            PathBuf::from("bin").join(hook_path.file_name().expect("copied hook binary filename")),
-            payload_path,
-        )
-    }
-
-    fn install_executable_script(&self, relative_path: &str, body: &str) -> PathBuf {
-        let path = self.tempdir.path().join(relative_path);
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent).expect("script dir");
-        }
-        fs::write(&path, body).expect("write script");
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-
-            let mut permissions = fs::metadata(&path).expect("script metadata").permissions();
-            permissions.set_mode(0o755);
-            fs::set_permissions(&path, permissions).expect("script permissions");
-        }
-        path
-    }
-
-    fn stdout(&self, output: &std::process::Output) -> String {
-        String::from_utf8(output.stdout.clone()).expect("stdout utf8")
-    }
-
-    fn stdout_json(&self, output: &std::process::Output) -> serde_json::Value {
-        serde_json::from_slice(&output.stdout).expect("valid json")
-    }
-
-    fn stderr(&self, output: &std::process::Output) -> String {
-        String::from_utf8(output.stderr.clone()).expect("stderr utf8")
+    fn deref(&self) -> &Self::Target {
+        &self.0
     }
 }
